--story=1003277 --user=赖文韬 保存流程评分设置bug https://www.tapd.cn/54247054/s/1086107

+++ src/main/java/codedriver/module/process/service/ProcessServiceImpl.java
@@ -143,7 +143,7 @@ public class ProcessServiceImpl implements ProcessService {
 		}
 
 		/** 保存评分设置 */
-		if(processVo.getProcessScoreTemplateVo() != null){
+		if(processVo.getProcessScoreTemplateVo() != null && processVo.getProcessScoreTemplateVo().getIsActive() == 1){
 			scoreTemplateMapper.deleteProcessScoreTemplateByProcessUuid(processVo.getUuid());
 			scoreTemplateMapper.insertProcessScoreTemplate(processVo.getProcessScoreTemplateVo());
 		}
